--- a/index.js
+++ b/index.js
@@ -1,11 +1,6 @@
-module.exports = function(n) {
-  if (n < 0) {
-    throw new RangeError("Cannot take factorial out of a negative number.");
-  } else {
-    return factorial(n);
-  }
-}
-
-function factorial(n) {
-  return n === 0 ? 1 : n * factorial(n - 1);
+module.exports = function factorial(n) {
+  if (n === 0) return 1
+  if (n < 0) return NaN
+  for (var r = 1; n > 1; n--) r *= n
+  return r
 }