hi, decided to try speeding this up a bit and got some pretty good results:
```js
# recursive(21) 10000000 times
ok ~2.78 s (2 s + 784540567 ns)

# imperative(21) 10000000 times
ok ~411 ms (0 s + 411079554 ns)
```
new version uses a simple `for` loop instead of recursion and is ~7x faster as a result:tada:

also, for the sake of mathematical accuracy and ease of use, i decided to return `NaN` for negative numbers instead of throwing an error. thoughts?