@contra 2017-06-24
What version of node are you testing this on? If you have something with tail call optimization I'd be curious to see the benchmarks.