@semibran 2017-06-24
looks like I was using `v8.0.0` last time. Upgraded to `v8.1.2` and got this:
```
NANOBENCH version 2
> /home/brandon/.nvm/versions/node/v8.1.2/bin/node test/bench
```
```js
# recursive(21) 10000000 times
ok ~2.83 s (2 s + 826264609 ns)

# imperative(21) 10000000 times
ok ~429 ms (0 s + 428565140 ns)
```