@contra 2017-06-24
Can you include a benchmark with a really large number? If it beats it there I'll merge it.