@semibran 2017-06-24
`factorial(170)` happens to be the largest number that doesn't evaluate to infinity, so i tried that.
```js
# recursive(170) 1000000 times
ok ~2.28 s (2 s + 284811122 ns)

# imperative(170) 1000000 times
ok ~478 ms (0 s + 477627853 ns)
```