@contra 2017-06-24
Oops, this makes the tests fail - can you fix?